Replaced implicit conversion of vector types with explicit conversion.

+++ src/gs/gsTile.cpp
@@ -139,8 +139,7 @@ GLuint gs::Tile::SetBufferOffset( const GLuint vertexCount )
 gs::Tile::Tile( const Type surface, const vector<shared_ptr<gs::Vertex>>& vertices, const gs::Vec3f& centroid, const double height )
     :   id( idCounter++ ),
         surface( surface ),
-        //centroid( centroid ),
-        normal( gs::Vec3f( centroid.x, centroid.y, centroid.z ).Unit() ),
+        normal( gs::Vec3d( (double) centroid.x, (double) centroid.y, (double) centroid.z ).Unit() ),
         height( height ),
         vertices( vertices ),
         fog( false )
